Spotify Integration + Database changes
A new class file called songdata where we can get song data for songs we are finger printing and adding to our database. Info is obtained from acoustid and spotify. Spotify mainly due to a very nice API and the ability to tell if a song is explicit or not. very important for me