GPU Support for Peak Finding & Spectograms
Finally added drop-in support for Peak Finding and Spectrogram generation. This should work with the existing database if you have it and have not changed setting related to defining peaks.

Some other print statements were deleted, GPU.py was completely rewritten, and moved the numpy - operator to ^ since it has been deprecated in new releases